requirements/pip.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-decouple==2.1
django-filter==2.2.0
django-rest-passwordreset==1.1.0
django-storages==1.9.1
djangorestframework==3.11.0
idna==2.10
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
sqlparse==0.4.1
urllib3==1.25.10
